Move KeycloakClient from shared-web to main application
