+++ cosmetics-web/app/models/user.rb
@@ -22,7 +22,7 @@ class User < ActiveHash::Base
 
   def self.load(force: false)
     begin
-      self.data = Shared::Web::KeycloakClient.instance.all_users(force: force)
+      self.data = KeycloakClient.instance.all_users(force: force)
     rescue StandardError => e
       Rails.logger.error "Failed to fetch users from Keycloak: #{e.message}"
       self.data = nil
