+++ shared-web/app/helpers/shared/web/user_management_helper.rb
@@ -2,7 +2,7 @@ module Shared
   module Web
     module UserManagementHelper
       def user_account_url
-        Shared::Web::KeycloakClient.instance.user_account_url
+        ::KeycloakClient.instance.user_account_url
       end
     end
   end
